+++ modules/actor/actor-factory.js
@@ -30,7 +30,7 @@ function factory(entities, baseClass) {
       const constructor = entities[data.type];
       if (!constructor) {
         // emit error
-        const error = `${SystemUtils.Localize("CPR.unsupportedentityerror")}: ${data.type}`;
+        const error = `${SystemUtils.Localize("CPR.messages.unsupportedEntityError")}: ${data.type}`;
         SystemUtils.DisplayMessage("error", error);
         throw new Error(error);
       }
@@ -44,7 +44,7 @@ function factory(entities, baseClass) {
           return (data, options) => {
             const constructor = entities[data.type];
             if (!constructor) {
-              const error = `${SystemUtils.Localize("CPR.unsupportedentityerror")}: ${data.type}`;
+              const error = `${SystemUtils.Localize("CPR.messages.unsupportedEntityError")}: ${data.type}`;
               SystemUtils.DisplayMessage("error", error);
               throw new Error(error);
             }