+++ modules/actor/sheet/cpr-container-sheet.js
@@ -1,11 +1,15 @@
-/* global mergeObject $ game getProperty */
+/* global mergeObject $ game getProperty duplicate */
 import CPRActorSheet from "./cpr-actor-sheet.js";
 import LOGGER from "../../utils/cpr-logger.js";
 import SystemUtils from "../../utils/cpr-systemUtils.js";
 import CPRChat from "../../chat/cpr-chat.js";
+import CPRItem from "../../item/cpr-item.js";
+import PurchasePartPrompt from "../../dialog/cpr-purchase-part-prompt.js";
 
 /**
- * Implement the sheet for containers and shop keepers.
+ * Implement the sheet for containers and shop keepers. This extends CPRActorSheet to make use
+ * of owned-item management methods like _getOwnedItem and _deleteOwnedItem.
+ *
  * @extends {CPRActorSheet}
  */
 export default class CPRContainerActorSheet extends CPRActorSheet {
@@ -59,6 +63,8 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
    * @param {*} html - the DOM object
    */
   activateListeners(html) {
+    LOGGER.trace("activateListeners | CPRContainerSheet | Called.");
+
     // Selection of trade partner
     html.find(".trade-with-dropdown").change((event) => this._setTradePartner(event));
     // Create item in inventory
@@ -130,7 +136,11 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
           break;
         }
         case "purchase": {
-          await this._purchaseItem(item);
+          await this._purchaseItem(item, true);
+          break;
+        }
+        case "purchaseFraction": {
+          this._purchaseItem(item, false);
           break;
         }
         default: {
@@ -189,33 +199,66 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
    *
    * @private
    * @param {Item} item - object to be purchased
+   * @param {boolean} all - Toggle to purchase all of the items in the stack or just a part of them
    */
-  async _purchaseItem(item) {
+  async _purchaseItem(item, all) {
     LOGGER.trace("_purchaseItem | CPRContainerSheet | Called.");
     if (this.tradePartnerId === undefined || this.tradePartnerId === "") {
-      SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.tradewithwarn"));
+      SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.messages.tradeWithWarn"));
       return;
     }
-
-    const tradePartnerActor = game.actors.get(this.tradePartnerId);
-    if (!getProperty(this.actor.data, "flags.cyberpunk-red-core.items-free")) {
-      const price = item.data.data.price.market;
-      if (tradePartnerActor.data.data.wealth.value < price) {
-        SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.tradepricewarn"));
+    const transferredItemData = duplicate(item.data);
+    let cost = 0;
+    if (item.type === "ammo" && item.data.data.variety !== "grenade" && item.data.data.variety !== "rocket") {
+      // Ammunition, which is neither grenades nor rockets, are prices are for 10 of them (pg. 344)
+      cost = item.data.data.price.market / 10;
+    } else {
+      cost = item.data.data.price.market;
+    }
+    if (!all) {
+      const itemText = SystemUtils.Format("CPR.dialog.purchasePart.text", { itemName: item.name });
+      const formData = await PurchasePartPrompt.RenderPrompt(itemText).catch((err) => LOGGER.debug(err));
+      if (formData === undefined) {
         return;
       }
-      const { amount } = item.data.data;
+      const newAmount = parseInt(formData.purchaseAmount, 10);
+      if (newAmount < 1 || newAmount >= parseInt(item.data.data.amount, 10)) {
+        SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.dialog.purchasePart.wrongAmountWarning"));
+        return;
+      }
+      transferredItemData.data.amount = newAmount;
+      cost *= newAmount;
+    } else {
+      cost *= item.data.data.amount;
+    }
+    const tradePartnerActor = game.actors.get(this.tradePartnerId);
+    if (!getProperty(this.actor.data, "flags.cyberpunk-red-core.items-free")) {
+      if (tradePartnerActor.data.data.wealth.value < cost) {
+        SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.messages.tradePriceWarn"));
+        return;
+      }
+      const { amount } = transferredItemData.data;
+      const username = game.user.name;
       let reason = "";
       if (amount > 1) {
-        reason = `${SystemUtils.Format("CPR.tradelogmultiple", { amount, name: item.name, price })} - ${game.user.name}`;
+        reason = `${SystemUtils.Format("CPR.containerSheet.tradeLog.multiple",
+          { amount, name: item.name, price: cost })} - ${username}`;
       } else {
-        reason = `${SystemUtils.Format("CPR.tradelogsingle", { name: item.name, price })} - ${game.user.name}`;
+        reason = `${SystemUtils.Format("CPR.containerSheet.tradeLog.single",
+          { name: item.name, price: cost })} - ${username}`;
       }
-      await tradePartnerActor.deltaLedgerProperty("wealth", -1 * price, reason);
+      await tradePartnerActor.deltaLedgerProperty("wealth", -1 * cost, reason);
     }
-    await tradePartnerActor.createEmbeddedDocuments("Item", [item.data]);
+    if (tradePartnerActor.automaticallyStackItems(new CPRItem(transferredItemData))) {
+      await tradePartnerActor.createEmbeddedDocuments("Item", [transferredItemData]);
+    }
     if (!getProperty(this.actor.data, "flags.cyberpunk-red-core.infinite-stock")) {
-      await this._deleteOwnedItem(item, true);
+      if (all) {
+        await this._deleteOwnedItem(item, true);
+      } else {
+        const keepAmount = item.data.data.amount - transferredItemData.data.amount;
+        await this.actor.updateEmbeddedDocuments("Item", [{ _id: item.id, "data.amount": keepAmount }]);
+      }
     }
   }
 
@@ -233,7 +276,7 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
     const itemTypeNice = itemType.toLowerCase().capitalize();
     const itemString = "ITEM.Type";
     const itemTypeLocal = itemString.concat(itemTypeNice);
-    const itemName = `${SystemUtils.Localize("CPR.new")} ${SystemUtils.Localize(itemTypeLocal)}`;
+    const itemName = `${SystemUtils.Localize("CPR.actorSheets.commonActions.new")} ${SystemUtils.Localize(itemTypeLocal)}`;
     const itemImage = SystemUtils.GetDefaultImage("Item", itemType);
     const itemData = { img: itemImage, name: itemName, type: itemType };
     await this.actor.createEmbeddedDocuments("Item", [itemData]);
@@ -271,26 +314,33 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
     await this.actor.setFlag("cyberpunk-red-core", "container-type", containerType);
     switch (containerType) {
       case "shop": {
-        await this.actor.unsetFlag("cyberpunk-red-core", "items-free");
-        await this.actor.unsetFlag("cyberpunk-red-core", "players-create");
-        await this.actor.unsetFlag("cyberpunk-red-core", "players-delete");
-        await this.actor.unsetFlag("cyberpunk-red-core", "players-modify");
+        // setting flags is an async operation; we wait for them all in parallel
+        await Promise.all([
+          this.actor.unsetFlag("cyberpunk-red-core", "items-free"),
+          this.actor.unsetFlag("cyberpunk-red-core", "players-create"),
+          this.actor.unsetFlag("cyberpunk-red-core", "players-delete"),
+          this.actor.unsetFlag("cyberpunk-red-core", "players-modify"),
+        ]);
         break;
       }
       case "loot": {
-        await this.actor.unsetFlag("cyberpunk-red-core", "infinite-stock");
-        await this.actor.setFlag("cyberpunk-red-core", "items-free", true);
-        await this.actor.unsetFlag("cyberpunk-red-core", "players-create");
-        await this.actor.unsetFlag("cyberpunk-red-core", "players-delete");
-        await this.actor.unsetFlag("cyberpunk-red-core", "players-modify");
+        await Promise.all([
+          this.actor.unsetFlag("cyberpunk-red-core", "infinite-stock"),
+          this.actor.setFlag("cyberpunk-red-core", "items-free", true),
+          this.actor.unsetFlag("cyberpunk-red-core", "players-create"),
+          this.actor.unsetFlag("cyberpunk-red-core", "players-delete"),
+          this.actor.unsetFlag("cyberpunk-red-core", "players-modify"),
+        ]);
         break;
       }
       case "stash": {
-        await this.actor.unsetFlag("cyberpunk-red-core", "infinite-stock");
-        await this.actor.setFlag("cyberpunk-red-core", "items-free", true);
-        await this.actor.setFlag("cyberpunk-red-core", "players-create", true);
-        await this.actor.setFlag("cyberpunk-red-core", "players-delete", true);
-        await this.actor.setFlag("cyberpunk-red-core", "players-modify", true);
+        await Promise.all([
+          this.actor.unsetFlag("cyberpunk-red-core", "infinite-stock"),
+          this.actor.setFlag("cyberpunk-red-core", "items-free", true),
+          this.actor.setFlag("cyberpunk-red-core", "players-create", true),
+          this.actor.setFlag("cyberpunk-red-core", "players-delete", true),
+          this.actor.setFlag("cyberpunk-red-core", "players-modify", true),
+        ]);
         break;
       }
       case "custom": {
@@ -302,12 +352,21 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
     }
   }
 
+  /**
+   * _onDrop is provided by ActorSheet and extended here. When an Item is dragged to an ActorSheet a new
+   * copy is created. This extension ensures players are allowed to create items in the container-actor before doing so.
+   *
+   * @private
+   * @override
+   * @param {Object} event - an object capturing event details
+   */
   async _onDrop(event) {
+    LOGGER.trace("_onDrop | CPRContainerSheet | Called.");
     const playersCanCreate = getProperty(this.actor.data, "flags.cyberpunk-red-core.players-create");
     if (game.user.isGM || playersCanCreate) {
       super._onDrop(event);
     } else {
-      SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.tradedraginwarn"));
+      SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.messages.tradeDragInWarn"));
     }
   }
 }